requirements.txt
openai>=1.83.0
python-dotenv
pydantic